fix(build) pin bootstrap dep (#136)
* fix(build) pin bootstrap dep

* fix(test) introduce pytest ordering to prioritize healthy test before the rest of the test suite
