+++ test/molecule-role/molecule/kubernetes/tests/test_healthy.py
@@ -1,6 +1,6 @@
 import os
 import util
-
+import pytest
 from testinfra.utils.ansible_runner import AnsibleRunner
 
 testinfra_hosts = AnsibleRunner(os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('kubernetes-cluster-agent')
@@ -8,6 +8,7 @@ testinfra_hosts = AnsibleRunner(os.environ['MOLECULE_INVENTORY_FILE']).get_hosts
 kubeconfig_env = "KUBECONFIG=/home/ubuntu/deployment/aws-eks/tf-cluster/kubeconfig "
 
 
+@pytest.mark.first
 def test_receiver_healthy(host):
     def assert_healthy():
         c = "curl -s -o /dev/null -w \"%{http_code}\" http://localhost:7077/health"
@@ -16,6 +17,7 @@ def test_receiver_healthy(host):
     util.wait_until(assert_healthy, 30, 5)
 
 
+@pytest.mark.second
 def test_node_agent_healthy(host):
     def assert_healthy():
         c = kubeconfig_env + "kubectl wait --for=condition=ready --timeout=1s -l app=stackstate-agent pod"
@@ -24,6 +26,7 @@ def test_node_agent_healthy(host):
     util.wait_until(assert_healthy, 30, 5)
 
 
+@pytest.mark.third
 def test_cluster_agent_healthy(host):
     def assert_healthy():
         c = kubeconfig_env + "kubectl wait --for=condition=available --timeout=1s deployment/stackstate-cluster-agent"
