+++ test/molecule-role/requirements.txt
@@ -58,6 +58,7 @@ pycparser==2.19
 pyflakes==2.1.1
 PyNaCl==1.3.0
 pytest==4.4.1
+pytest-ordering==0.6
 python-dateutil==2.8.0
 python-gilt==1.2.1
 pywinrm==0.3.0
